docs(readme): add link to component collection (#31)
I was looking for spinning loader animations and shared this library as components using bit (https://github.com/teambit/bit).
This is quite useful as now people can:

1) Discover each of the components via bit’s platform and Google
2) Install each component individually with npm and yarn
3) use bit to import the component and develop them in their own projects - which helps as people can easily change and style the components as needed.
4) Play with each component in an online playground.

I’ve added the link to the readme. I’d be happy to maintain the collection, and/or hand over ownership if you prefer (---).
Hope you like it. Cheers